+++ tlsn-core/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "notarized-session"
+name = "tlsn-core"
 version = "0.1.0"
 edition = "2021"
