+++ tlsn-core/src/lib.rs
@@ -6,8 +6,8 @@ pub mod session_data;
 pub mod session_header;
 pub mod transcript;
 
-use session_data::SessionData;
-use session_header::SessionHeader;
+pub use session_data::SessionData;
+pub use session_header::SessionHeader;
 
 pub type HashCommitment = [u8; 32];
 
@@ -17,16 +17,16 @@ pub type LabelSeed = [u8; 32];
 
 pub struct NotarizedSession {
     version: u8,
-    session_header: SessionHeader,
-    session_data: SessionData,
+    header: SessionHeader,
+    data: SessionData,
 }
 
 impl NotarizedSession {
-    pub fn new(version: u8, session_header: SessionHeader, session_data: SessionData) -> Self {
+    pub fn new(version: u8, header: SessionHeader, data: SessionData) -> Self {
         Self {
             version,
-            session_header,
-            session_data,
+            header,
+            data,
         }
     }
 
@@ -34,11 +34,11 @@ impl NotarizedSession {
         self.version
     }
 
-    pub fn session_header(&self) -> &SessionHeader {
-        &self.session_header
+    pub fn header(&self) -> &SessionHeader {
+        &self.header
     }
 
-    pub fn session_data(&self) -> &SessionData {
-        &self.session_data
+    pub fn data(&self) -> &SessionData {
+        &self.data
     }
 }
